face_emotion_analysis_service/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-multipart==0.0.6
motor==3.2.0
pymongo==4.5.0
python-dotenv==1.0.0
deepface==0.0.78
gunicorn